Upgrade to React 16 and lint - Refactor, split component and containers
