+++ src/components/Main.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import HistoryList from './HistoryList';
-import SocketPlayground from './SocketPlayground';
+import HistoryList from '../containers/HistoryList';
+import SocketPlayground from '../containers/SocketPlayground';
 import SocketPlaygroundBar from './SocketPlaygroundBar';
 import Row from './Row';
 import Column from './Column';
